Fix: Show all plugins for subrecord selection
The code was only displaying the plugins that contained an entry for the
subrecord the first time it was encountered.  This resulted in the user
being prompted to select from a single plugin on occassions and did not
allow them to select a plugin that may only show up in a later record.